+++ src/main/java/com/example/tenpaws/domain/facility/entity/Facility.java
@@ -1,0 +1,19 @@
+package com.example.tenpaws.domain.facility.entity;
+
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import org.springframework.data.annotation.Id;
+
+@Entity
+public class Facility {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private String name;
+    private Point location;
+
+    // Getters and Setters
+}
