+++ src/main/java/com/example/tenpaws/domain/shelter/repository/ShelterRepository.java
@@ -1,0 +1,9 @@
+package com.example.tenpaws.domain.shelter.repository;
+
+import com.example.tenpaws.domain.shelter.entity.Shelter;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ShelterRepository extends JpaRepository<Shelter, Long> {
+}
