--- src/main/java/com/example/tenpaws/facility/entity/Facility.java
@@ -1,19 +1,0 @@
-package com.example.tenpaws.facility.entity;
-
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import org.springframework.data.annotation.Id;
-
-@Entity
-public class Facility {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String name;
-    private Point location;
-
-    // Getters and Setters
-}
