--- src/main/java/com/example/tenpaws/facility/service/FacilityService.java
@@ -1,31 +1,0 @@
-package com.example.tenpaws.facility.service;
-
-import com.example.tenpaws.facility.Facility;
-import com.example.tenpaws.facility.repository.FacilityRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-public class FacilityService {
-
-    @Autowired
-    private FacilityRepository facilityRepository;
-
-    public List<FacilityDTO> getNearbyFacilities(Double lat, Double lng) {
-        return facilityRepository.findByLocation(lat, lng).stream()
-                .map(this::toFacilityDTO)
-                .collect(Collectors.toList());
-    }
-
-    private FacilityDTO toFacilityDTO(Facility facility) {
-        FacilityDTO dto = new FacilityDTO();
-        dto.setId(facility.getId());
-        dto.setName(facility.getName());
-        dto.setLat(facility.getLat());
-        dto.setLng(facility.getLng());
-        return dto;
-    }
-}
