--- src/main/java/com/example/tenpaws/pet/dto/PetRequestDTO.java
@@ -1,13 +1,0 @@
-package com.example.tenpaws.pet.dto;
-
-import lombok.Data;
-
-@Data
-public class PetRequestDTO {
-    private String species;
-    private String size;
-    private int age;
-    private String personality;
-    private int exerciseLevel;
-}
-
