--- src/main/java/com/example/tenpaws/pet/dto/PetResponseDTO.java
@@ -1,13 +1,0 @@
-package com.example.tenpaws.pet.dto;
-
-import lombok.Data;
-
-@Data
-public class PetResponseDTO {
-    private Long petId;
-    private String species;
-    private String size;
-    private int age;
-    private String personality;
-    private int exerciseLevel;
-}
