--- src/main/java/com/example/tenpaws/shelter/repository/ShelterRepository.java
@@ -1,9 +1,0 @@
-package com.example.tenpaws.shelter.repository;
-
-import com.example.tenpaws.shelter.entity.Shelter;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ShelterRepository extends JpaRepository<Shelter, Long> {
-}
