+++ src/main/resources/application.properties
@@ -10,5 +10,6 @@ spring.jpa.properties.hibernate.show_sql=true
 spring.jpa.properties.hibernate.format_sql=true
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQLDialect
 
+spring.jwt.secret=${SPRING_SECRET_KEY:vmfhaltmskdlstkfkdgodyroqkfwkdbalroqkfwkdbalaaaaaaaaaaaaaaaabbbbb}
 
 logging.level.org.hibernate.sql=debug
